fix(v3): restore model.to(bf16) — Triton PLIF kernels require bf16 alignment
Removed .to(bfloat16) assuming DS bf16 would handle it; instead got
CUDA misaligned address at line 1640 forward_parallel (Triton kernels
assume bf16 pointer alignment).

Keep model.to(bf16) + ds_config.bf16.enabled=true. DS upcasts bf16 params
to fp32 master on engine init, providing proper mixed precision.

+++ scripts/test_muon_ds.py
@@ -60,9 +60,9 @@ def main():
             D=128, N=2, K=6, num_layers=4, D_ff=256,
             vocab_size=256, memory_layer_interval=2,
         )
-    # Keep params fp32; DS bf16.enabled=true manages mixed precision properly
-    # (bf16 compute + fp32 master copy + fp32 optimizer state).
-    m = NeuronSparkForCausalLM(cfg).cuda()
+    # Params must be bf16 for Triton PLIF kernel alignment (see train_pretrain.py).
+    # DS bf16.enabled=true on top gives fp32 master copy + fp32 Adam state.
+    m = NeuronSparkForCausalLM(cfg).cuda().to(torch.bfloat16)
     promote_neuron_params_fp32(m)
 
     if args.optimizer == "muon":
